+++ src/components/general/common.module.css
@@ -1,4 +1,3 @@
-/*todo need to rename this and figure out its purpose*/
 
 .row {
     display: flex;
@@ -128,3 +127,8 @@
     opacity: 1;
     transition: 0s;
 }
+
+.fullscreenHeight {
+    min-height: 100vh;
+}
+
