+++ src/components/login/Login.tsx
@@ -13,7 +13,7 @@ import {
 } from "../../slices/bankingSlice";
 import {useAppSelector} from "../../app/hooks";
 import CustomTextField from "../general/CustomTextField";
-import styles from "../banking/Banking.module.css";
+import styles from "../general/common.module.css";
 import SimpleSnackbar from "../general/SimpleSnackbar";
 import logo from "../../gojenga1.webp";
